added gguf

+++ utils/file_utils.py
@@ -70,23 +70,25 @@ def oai_get_batch_res(client, pred_path=os.path.join("files", "preds")):
             merged_res = []
             path_to_file = os.path.join(pred_path, filename[0])
 
-            data = []
-            with open(path_to_file, "r") as f:
-                for line in f:
-                    data.append(json.loads(line.strip()))
-            batch_res = client.files.content(batch.output_file_id).text
-            batch_res = [json.loads(line) for line in batch_res.splitlines()]
+            if os.path.exists(path_to_file):
+                data = []
+                
+                with open(path_to_file, "r") as f:
+                    for line in f:
+                        data.append(json.loads(line.strip()))
+                batch_res = client.files.content(batch.output_file_id).text
+                batch_res = [json.loads(line) for line in batch_res.splitlines()]
 
-            if pred_path == os.path.join("files", "preds"):
-                for sample in data:
-                    res = [res["response"]["body"]["choices"][0]["message"]["content"] for res in batch_res if res["custom_id"] == sample["custom_id"]]
-                    merged_res.append({
-                        "id": sample["custom_id"],
-                        "prompt": sample["body"]["messages"][0]["content"],
-                        "output": res[0].strip(),
-                        "model_inf_time": "n/a", 
-                })
-                with open(os.path.join(pred_path, f"{filename[0].split('.')[0]}.json"), "w") as f:
-                    json.dump({
-                        "golds": merged_res
-                    }, f)
+                if pred_path == os.path.join("files", "preds"):
+                    for sample in data:
+                        res = [res["response"]["body"]["choices"][0]["message"]["content"] for res in batch_res if res["custom_id"] == sample["custom_id"]]
+                        merged_res.append({
+                            "id": sample["custom_id"],
+                            "prompt": sample["body"]["messages"][0]["content"],
+                            "output": res[0].strip(),
+                            "model_inf_time": "n/a", 
+                    })
+                    with open(os.path.join(pred_path, f"{filename[0].split('.')[0]}.json"), "w") as f:
+                        json.dump({
+                            "golds": merged_res
+                        }, f)
